Make CompletionHandler::run return an array instead of a string with new lines
Completion results are handled inside CompletionHandler as arrays, but were
being converted to a string on output as this is the format BASH completion
uses. Since `OutputInterface::write` happily accepts an array of strings and
effectively outputs the same format to the host completion system, it makes
the internals nicer if CompletionHandler returns arrays of terms and leaves
output formatting to the OutputInterface.