Only run unit tests on travis with --prefer-lowest once
Per the comments in the thread for #34, I don't like the idea of running
tests twice for every PHP version (high and low dependency checks). It
is comprehensive, but it seems a little redunant and wasteful. One check
should be enough to pick up any general issues with the lowest version
dependencies, though some problems may slip through because of this. I'm
happy to have the test run take less time in exchange for this minor
risk:

> testing all combinations feels like testing that the dependencies work
with the PHP versions they claim to, more than it is testing the code in
this library.