Prevent global/application options conflicting with CompletionCommand
`CompletionCommand` does not need to know about custom application-level options
in user code. This filters out options that aren't in the default set provided
by the base `Application` class to prevent conflicts with option names while
maintaining backwards compatibility.

If user-code wants to use global options in a subclass of CompletionCommand,
`filterApplicationOptions` can be overridden to append any desired options to
the default list.

Solves the issue mentioned in #86