Escape completion results for BASH before writing to stdout
In the previous commit I used `printf '%q'` in the BASH hook to handle escaping
unquoted strings, however this didn't handle escaping quote characters inside
strings where needed. This commit moves the responsibility of escaping completion
results into PHP to make the escaping logic testable and easier to maintain.

BASH supports automatic escaping/quoting in completion results only when the
`-o filenames` option is used to register a completion function. The `filenames`
option has behaviour that is incompatible with a generic completion handler,
such as appending a `/` on results that happen to match directory name in the CWD.
Unfortuntately this means we cannot use this functionality. There is no API to
change this behaviour as far as I've been able to see from the BASH docs and
source code.

ZSH handles escaping of completion functions perfectly fine out of the box.