Use older HHVM version in Travis to fix build
The `hhvm` image in Travis has started using HHVM 4.x which appears to
have a problem with the composer version used (1.2.x):

    Fatal error: Uncaught Error: Keyword __HALT_COMPILER must be written
    in lowercase in /home/travis/.phpenv/versions/hhvm-stable/bin/composer:27

This changes the `prefer-lowest` HHVM build to use the oldest HHVM image
available (3.3) and adds the 4.x HHVM build as allowed to fail. This
isn't a permanent change - it's just to get the build passing again for
now.