Fix command not set correctly when options precede the command name
The CompletionHandler code made an assumption that the command name
was always the second word on the command line:

    $ [program-name] [command-name] [...arguments and options]

This now searches for the first non-option-like word and stores its
index for other parts of CompletionHandler to test against.

The way ArrayInput is built in `CompletionHandler::getInput()` doesn't
align with how the docs for that class expect it to be used, so calling
`getFirstArgument()` was always returning the first value.

Fixes #83